implementing a list view graphlet is not require

+++ DataAccess/sqlite3/vsqlite3.cpp
@@ -102,7 +102,7 @@ std::string VirtualSQLite3::insert_into(const char* tablename, bool replace) {
 	return sql;
 }
 
-std::string VirtualSQLite3::select_from(const char* tablename, const char* order_by, unsigned int limit, unsigned int offset) {
+std::string VirtualSQLite3::select_from(const char* tablename, const char* order_by, uint64 limit, uint64 offset) {
 	std::string sql = columns_join("SELECT ", ", ", " ", this->columns, this->count);
 	std::string maybe_order_by = ((order_by == nullptr)? " " : make_nstring(" ORDER BY %s ", order_by));
 
@@ -112,7 +112,7 @@ std::string VirtualSQLite3::select_from(const char* tablename, const char* order
 	return sql;
 }
 
-std::string VirtualSQLite3::select_from(const char* tablename, const char* order_by, const char* cols[], size_t count, unsigned int limit, unsigned int offset) {
+std::string VirtualSQLite3::select_from(const char* tablename, const char* order_by, const char* cols[], size_t count, uint64 limit, uint64 offset) {
 	std::string sql = columns_join("SELECT ", ", ", " ", cols, count);
 	std::string maybe_order_by = ((order_by == nullptr) ? " " : make_nstring(" ORDER BY %s ", order_by));
